requirements.txt
# Solray AI — Backend Requirements
# Production-ready for Railway + Supabase PostgreSQL

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (async SQLAlchemy + SQLite for local dev, PostgreSQL for prod)
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4

# HTTP / Input validation
python-multipart==0.0.9
pydantic[email]==2.7.1
email-validator==2.1.1

# Astrology engines
pyswisseph==2.10.3.2
geopy==2.4.1

# AI
anthropic==0.25.0

# Dev utilities
python-dotenv==1.0.1
httpx==0.27.0
